feat: stabilize pipeline with IDW gap-limiting, recursive fix, and visual workflow stepper
- Fix infinite recursion between GEEClient.fetch_in_chunks and provider.get_rainfall_data via _is_chunk flag
- Fix point sampling API crash by adding GEE auth check to /api/jobs/point endpoint
- Fix Pandas 2.0 Timestamp.value deprecation in client.py error handler
- Add strict daily reindexing in processor.py to fill temporal gaps with NaN
- Add max_gap=3 to IDW interpolator to prevent fabricating data over large missing blocks
- Replace BackgroundTasks with daemon threads for instant Uvicorn hot-reload
- Add interactive visual pipeline workflow stepper (HTML/CSS/JS) with 6 animated stages
- Update test suite: mock signatures, provider assertions, exporter patches (41 tests pass)

+++ src/pipeline/analytics.py
@@ -16,8 +16,8 @@ def calculate_trend(df: pd.DataFrame, column: str = 'precipitation', min_points:
             'slope': 0.0
         }
 
-    # Prepare data (clean NaNs)
-    valid_data = df[column].dropna()
+    # Prepare data (clean NaNs and ensure numeric float type)
+    valid_data = df[column].dropna().astype(float)
     if len(valid_data) < min_points:
         return {
             'status': 'insufficient_data',
